D-40 Finished Sub Goals & Refactored Controllers
