+++ components/DashboardComponents/SubGoalsContainer/SubGoalsContainer.js
@@ -5,7 +5,7 @@ import ProgressionContainer from "./components/ProgressionContainer";
 import SubGoalsClient from "../../../util/clients/subGoalsClient";
 
 const SubGoalsContainer = () => {
-  const [taskAdded, setTaskAdded] = useState(true);
+  const [taskUpdated, setTaskUpdated] = useState(true);
   const [stages, setStages] = useState([
     {
       text: "To-Do",
@@ -26,35 +26,42 @@ const SubGoalsContainer = () => {
       tasks: [],
     },
   ]);
+
+  function pushTaskIfNotExists(task, taskArray) {
+    // Check if the task already exists in the array
+    const taskExists = taskArray.some(subgoal => subgoal._id === task._id);
+  
+    // If the task doesn't exist, push it into the array
+    if (!taskExists) {
+      taskArray.push(task);
+    }
+  }
+
   const fetchData = async () => {
     try {
       const goals = await SubGoalsClient.getAllSubGoals();
-      goals.forEach (item => {
-        if (item.status === "To-do"){ 
-          stages[0].tasks = [];
-          stages[0].tasks.push(item);
+      goals.forEach(item => {
+        if (item.status === "To-do") {
+          pushTaskIfNotExists(item, stages[0].tasks);
+        } else if (item.status === "In progress") {
+          pushTaskIfNotExists(item, stages[1].tasks);
+        } else if (item.status === "Complete") {
+          pushTaskIfNotExists(item, stages[2].tasks);
         }
-        if (item.status === "In progress"){
-          stages[1].tasks = [];
-          stages[1].tasks.push(item)
-        }
-        if (item.status === "Complete"){
-          stages[2].tasks = [];
-          stages[2].tasks.push(item)
-        }
-        setTaskAdded(false)
-      })
+      });
+
+      setTaskUpdated(false)
   } catch (error) {
     console.log(error)
   }} 
 
-
+fetchData();
   
 useEffect (() => {
-  if (taskAdded){
-    fetchData()
-  } 
-}, [taskAdded])
+  if (!taskUpdated) {
+    fetchData();
+  }
+}, [taskUpdated])
 
   return (
     <div className="bg-opacity-30 w-full h-auto rounded-b-lg p-10 bg-neutral-100 shadow-md">
@@ -72,7 +79,7 @@ useEffect (() => {
             color={section.color}
             circleColor={section.circleColor}
             tasks={section.tasks}
-            setTaskAdded={setTaskAdded}
+            setTaskUpdated={setTaskUpdated}
           />
         ))}
       </div>
