+++ components/DashboardComponents/SubGoalsContainer/components/SmallSubGoals.js
@@ -1,12 +1,8 @@
 const SmallSubGoals = ({ task, setTaskClicked, setIsEditModalVisible }) => {
-
-
-  
-  
   // Opens section to edit Sub Goals
   const openModal = () => {
     setIsEditModalVisible(true);
-    setTaskClicked({ task });
+    setTaskClicked(task);
   };
 
   // const deleteSubGoal =  async (task) => {
@@ -22,10 +18,8 @@ const SmallSubGoals = ({ task, setTaskClicked, setIsEditModalVisible }) => {
       onClick={openModal}
       className="border border-indigo-600 h-20 w-[97%] hover:scale-105 transition duration-200 rounded-lg bg-white hover:cursor-pointer p-2"
     >
-      
       <span className="text-lg p-1">{task.icon}</span>
       <h1 className="text-black text-center">{task.title}</h1>
-      
     </div>
   );
 };
