+++ server/controllers/mainGoalsControllers.js
@@ -1,66 +1,6 @@
 require("dotenv").config();
 const createError = require("http-errors");
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-const User = require("../models/user");
 const MainGoals = require("../models/maingoal");
-
-exports.signUp = async (req, res, next) => {
-  try {
-    const hashedPassword = await bcrypt.hash(req.body.password, 10);
-    const user = new User({
-      username: req.body.username,
-      email: req.body.email,
-      password: hashedPassword,
-    });
-    await user.save();
-    res.json({ message: "Sign up" });
-  } catch (error) {
-    return next(createError(500, error.message));
-  }
-};
-
-exports.signIn = async (req, res, next) => {
-  try {
-    const userEmail = req.body.email;
-    const user = await User.findOne({ email: userEmail });
-    if (!user) {
-      return next(createError(404, "The user not found"));
-    }
-    const result = await bcrypt.compare(req.body.password, user.password);
-
-    if (result) {
-      const accessToken = jwt.sign(
-        { id: user._id },
-        process.env.ACCESS_TOKEN_SECRET
-      );
-      res.cookie("jwt", accessToken, {
-        httpOnly: true,
-        maxAge: 24 * 60 * 60 * 1000,
-      });
-    }
-    res.json({ message: "Sign in" });
-  } catch (error) {
-    return next(createError(500, error.message));
-  }
-};
-
-exports.getUser = async (req, res, next) => {
-  try {
-    res.json({ username: req.user.username, email: req.user.email });
-  } catch (error) {
-    return next(createError(500, error.message));
-  }
-};
-
-exports.signOut = async (req, res, next) => {
-  try {
-    res.clearCookie("jwt");
-    res.json({ message: "Sign out" });
-  } catch (error) {
-    return next(createError(500, error.message));
-  }
-};
 
 exports.addOneGoal = async (req, res, next) => {
   try {
@@ -141,20 +81,6 @@ exports.deleteOneMainGoal = async (req, res, next) => {
     return next(createError(500, error.message));
   }
 };
-
-
-exports.getSubGoals = async (req, res, next) => {
-  try {
-      const userId = req.user._id.toString();
-      const mainGoals = await MainGoals.find({ userId: userId });
-      res.json(mainGoals);
-      if (!mainGoals){
-          return next(createError(404, "No goals found"));
-      }
-  } catch (error) {
-      return next(createError(500, error.message));
-  }
-}
 
 exports.getAllMainGoalsForDashboard = async (req, res, next) => {
   try {
